+++ qiskit/aqua/circuits/__init__.py
@@ -15,7 +15,7 @@
 # limitations under the License.
 # =============================================================================
 
-from .boolean_logic_circuits import CNF, DNF, ESOP
+from .boolean_logical_circuits import CNF, DNF, ESOP
 from .phase_estimation_circuit import PhaseEstimationCircuit
 from .statevector_circuit import StateVectorCircuit
 from .fourier_transform_circuits import FourierTransformCircuits